requirements.txt
ffpyplayer
bilibili-api-python
aiohttp
fake-useragent